requirements.txt
Django==3.2.5
django-cors-headers==3.7.0
djangorestframework==3.12.4
psycopg2-binary==2.8.6
PyJWT==2.1.0
typing-extensions==3.10.0.0
moviepy==1.0.3
SpeechRecognition==3.8.1